Remove an unused parameter from sqlite3DequoteExpr.  Fix another unrelated and harmless compiler warning.

+++ src/notify.c
@@ -230,9 +230,9 @@ void sqlite3ConnectionUnlocked(sqlite3 *db){
 
       sqlite3BeginBenignMalloc();
       assert( aArg==aDyn || (aDyn==0 && aArg==aStatic) );
-      assert( nArg<=ArraySize(aStatic) || aArg==aDyn );
-      if( (!aDyn && nArg==ArraySize(aStatic))
-       || (aDyn && nArg==(sqlite3DbMallocSize(db, aDyn)/sizeof(void*)))
+      assert( nArg<=(int)ArraySize(aStatic) || aArg==aDyn );
+      if( (!aDyn && nArg==(int)ArraySize(aStatic))
+       || (aDyn && nArg==(int)(sqlite3DbMallocSize(db, aDyn)/sizeof(void*)))
       ){
         /* The aArg[] array needs to grow. */
         void **pNew = (void **)sqlite3Malloc(nArg*sizeof(void *)*2);
@@ -307,4 +307,3 @@ void sqlite3ConnectionClosed(sqlite3 *db){
   leaveMutex();
 }
 #endif
-
